Move the navbar and layout components to their own layout folder
