+++ zdblog-web/src/pages/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { Layout } from "../components/layout";
+import { Layout } from "../components/layout/layout";
 
 // markup
 const IndexPage = () => {
